requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
sentence-transformers==2.2.2
faiss-cpu==1.7.4
psutil==5.9.4
pydantic==1.10.7